+++ hms_irc/settings.py
@@ -6,8 +6,10 @@ RABBIT_HOST = 'localhost'                   # Address of the server
 RABBIT_EXCHANGE = 'haum'                    # Name of the direct exchanger
 RABBIT_COMMAND_ROUTING_KEY = 'irc_command'  # Routing key for irc command
 
+PING_NAME = 'hms_irc'
+
 # List of routing keys to listen to
-RABBIT_ROUTING_KEYS = ['reddit', 'irc_debug']
+RABBIT_ROUTING_KEYS = ['reddit', 'irc_debug', 'ping']
 
 # IRC