+++ hms_irc/handlers/ping.py
@@ -7,16 +7,17 @@ def get_logger():
 
 def handle(irc_server, irc_chan, msg):
 
-    if msg['type'] == 'request':
+    if msg['type'] == 'request' and msg['source'] == 'irc':
 
-        privmsg = 'Une requêtre globale de ping a été envoyée {}'.format(msg)
+        privmsg = ('Ping envoyé. All your microservices are belong to '
+                   'us.'.format(msg))
 
         irc_server.privmsg(irc_chan, privmsg)
         get_logger().info('Posted ping message: {}'.format(privmsg))
 
-    elif msg['type'] == 'answer':
+    elif msg['type'] == 'answer' and msg['source']['source'] == 'irc':
 
-        privmsg = 'Réponse reçue au ping {}'.format(msg)
+        privmsg = 'Le service `{}` a répondu présent'.format(msg['name'])
 
         irc_server.privmsg(irc_chan, privmsg)
         get_logger().info('Posted pong message: {}'.format(privmsg))